hessian 2pl working for items

+++ src/pl2_est.cpp
@@ -55,7 +55,7 @@ void estep_pl2(field<mat>& itrace, const ivec& pni, const ivec& pcni, const ivec
 }
 
 // [[Rcpp::export]]
-double loglikelihood_2pl(const arma::imat& a, const arma::vec& A, const arma::mat& b, const arma::ivec& ncat, 
+long double loglikelihood_2pl(const arma::imat& a, const arma::vec& A, const arma::mat& b, const arma::ivec& ncat, 
 				const arma::ivec& pni, const arma::ivec& pcni, const arma::ivec& pi, const arma::ivec& px, 
 				const arma::vec& theta, const arma::vec& mu, const arma::vec& sigma, const arma::ivec& pgroup)
 {
@@ -87,8 +87,10 @@ double loglikelihood_2pl(const arma::imat& a, const arma::vec& A, const arma::ma
 			ll += std::log(accu(posterior)); 
 		}
 	}
-	return (double)ll;
+	return ll;
 }
+
+
 
 
 // stop estimation because of decreasing likelihood
